FhirPatientService test reorganization with phpunit parameterized tests
