+++ tests/Tests/Services/FHIR/FhirPatientServiceCrudTest.php
@@ -1,0 +1,86 @@
+<?php
+
+namespace OpenEMR\Tests\Services\FHIR;
+
+use PHPUnit\Framework\TestCase;
+use OpenEMR\Tests\Fixtures\FixtureManager;
+use OpenEMR\Services\FHIR\FhirPatientService;
+use OpenEMR\FHIR\R4\FHIRDomainResource\FHIRPatient;
+
+/**
+ * FHIR Patient Service Crud Tests
+ * @coversDefaultClass OpenEMR\Services\FHIR\FhirPatientService
+ * @package   OpenEMR
+ * @link      http://www.open-emr.org
+ * @author    Dixon Whitmire <dixonwh@gmail.com>
+ * @copyright Copyright (c) 2020 Dixon Whitmire <dixonwh@gmail.com>
+ * @license   https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
+ *
+ */
+class FhirPatientServiceCrudTest extends TestCase
+{
+    private $fixtureManager;
+    private $patientFixture;
+    private $fhirPatientFixture;
+    private $fhirPatientService;
+
+    protected function setUp(): void
+    {
+        $this->fixtureManager = new FixtureManager();
+        $this->patientFixture = (array) $this->fixtureManager->getSinglePatientFixture();
+        $this->fhirPatientFixture = (array) $this->fixtureManager->getSingleFhirPatientFixture();
+        $this->fhirPatientService = new FhirPatientService();
+    }
+
+    protected function tearDown(): void
+    {
+        $this->fixtureManager->removePatientFixtures();
+    }
+
+    /**
+     * Tests a successful insert operation
+     * @covers ::insert
+     * @covers ::insertOpenEMRRecord
+     */
+    public function testInsert()
+    {
+        $processingResult = $this->fhirPatientService->insert($this->fhirPatientFixture);
+        $this->assertTrue($processingResult->isValid());
+        $this->assertGreaterThan(0, $processingResult->getData()[0]['pid']);
+    }
+
+    /**
+     * Tests an insert operation where an error occurs
+     * @covers ::insert
+     * @covers ::insertOpenEMRRecord
+     */
+    public function testInsertWithErrors()
+    {
+        unset($this->fhirPatientFixture['name']);
+        $processingResult = $this->fhirPatientService->insert($this->fhirPatientFixture);
+        $this->assertFalse($processingResult->isValid());
+        $this->assertEquals(0, count($processingResult->getData()));
+    }
+
+    /**
+     * Tests a successful update operation
+     * @covers ::update
+     * @covers ::updateOpenEMRRecord
+     */
+    public function testUpdate()
+    {
+        $this->expectException(\Exception::class);
+        $this->fhirPatientService->update('a-fhir-resource-id', $this->fhirPatientFixture);
+    }
+
+    /**
+     * Tests an update operation where an error occurs
+     * @covers ::update
+     * @covers ::updateOpenEMRRecord
+     */
+    public function testUpdateWithErrors()
+    {
+        $this->expectException(\Exception::class);
+        $this->fhirPatientService->update('a-fhir-resource-id', $this->fhirPatientFixture);
+    }
+}
